+++ src/rpc/util.h
@@ -13,6 +13,7 @@
 #include <rpc/request.h>
 #include <script/script.h>
 #include <script/sign.h>
+#include <script/signingprovider.h>
 #include <script/standard.h>
 #include <univalue.h>
 #include <util/check.h>
