+++ src/script/standard.cpp
@@ -8,7 +8,6 @@
 #include <pubkey.h>
 #include <script/script.h>
 
-
 typedef std::vector<unsigned char> valtype;
 
 bool fAcceptDatacarrier = DEFAULT_ACCEPT_DATACARRIER;
