+++ src/test/denialofservice_tests.cpp
@@ -6,10 +6,11 @@
 
 #include <banman.h>
 #include <chainparams.h>
-#include <keystore.h>
 #include <net.h>
 #include <net_processing.h>
 #include <script/sign.h>
+#include <script/signingprovider.h>
+#include <script/standard.h>
 #include <util/memory.h>
 #include <util/system.h>
 #include <util/time.h>
