+++ src/test/script_tests.cpp
@@ -6,10 +6,10 @@
 
 #include <core_io.h>
 #include <key.h>
-#include <keystore.h>
 #include <script/script.h>
 #include <script/script_error.h>
 #include <script/sign.h>
+#include <script/signingprovider.h>
 #include <util/strencodings.h>
 #include <rpc/util.h>
 #include <test/util/setup_common.h>
