+++ src/test/transaction_tests.cpp
@@ -12,11 +12,11 @@
 #include <consensus/validation.h>
 #include <core_io.h>
 #include <key.h>
-#include <keystore.h>
 #include <validation.h>
 #include <policy/policy.h>
 #include <policy/settings.h>
 #include <script/script.h>
+#include <script/signingprovider.h>
 #include <script/script_error.h>
 #include <streams.h>
 #include <util/strencodings.h>
