+++ src/wallet/ismine.cpp
@@ -7,6 +7,7 @@
 
 #include <key.h>
 #include <script/script.h>
+#include <script/signingprovider.h>
 #include <wallet/wallet.h>
 
 typedef std::vector<unsigned char> valtype;
